added moxie pass check

+++ services/payflow-service/src/main/java/ua/sinaver/web3/payflow/service/NotificationService.java
@@ -313,7 +313,7 @@ public class NotificationService {
 			}
 		}
 
-		val tokenAmount = Integer.parseInt(payment.getTokenAmount());
+		val tokenAmount = Double.parseDouble(payment.getTokenAmount());
 		val tokenAmountText = tokenAmount > 1 ? tokenAmount + "x " : "";
 
 		String castText;
